backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
click==8.3.2
cmdstanpy==1.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.3
fonttools==4.62.1
greenlet==3.3.2
h11==0.16.0
holidays==0.93
idna==3.11
importlib_resources==6.5.2
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
lightgbm==4.6.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==2.2.6
# packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.49
stack-data==0.6.3
stanio==0.5.1
starlette==1.0.0
threadpoolctl==3.6.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.43.0
wcwidth==0.6.0
widgetsnbextension==4.0.15
